backend/requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.3
distlib==0.4.0
Django==5.2.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.15
filelock==3.20.1
Flask==3.1.2
Flask-Admin==2.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.11
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.8
packaging==26.0
pillow==11.2.1
platformdirs==4.5.0
pycparser==2.23
PyMySQL==1.1.2
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
qrcode==8.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.35.4
Werkzeug==3.1.4
WTForms==3.2.1
django-cloudinary-storage
cloudinary
djangorestframework-simplejwt
django-redis
celery